Android: use page to host multiple pages
In order to make room for a separate UI for testing the hazmat
functionality, refactor the main activity such that it is now a
ViewPager. That pager will later host not just the current fragment
with the test for the sss intermediate functionality but also a second
fragment with a test for the low level hazmat functionality.